// SisAdmConvAux/src/app/models/courts.model.ts
export interface Courts {
  description: string;
  meritsCourts: number;
  studentDelegateMerit: number;
  knowledgeCourts: number;
  studentDelegateKnowledge: number;
  idannouncement: number;
}

export interface Courts1 {
  description: string;
  meritsCourts: number;
  studentDelegateMerit: number;
  knowledgeCourts: number;
  studentDelegateKnowledge: number;
  idannouncement: number;
  numberMeritCourts: number;
  numberKnowledgeCourts: number;
  numberKnowledgeStudentDelegate: number;
  numberMeritStudentDelegate: number;
}